requirements.txt
streamlit==1.30.0
requests==2.31.0
cryptography==41.0.5
python-dotenv==1.0.0
snowflake-connector-python==3.4.0
